chore: script commands setup
